✨ feat: PhysicalInfoScreen에 AppleHealthKit 데이터 기본값 사용 추가
